refactor(orchestrator): extract config, solana, and util modules from main.rs
Extract three focused modules from the monolithic main.rs:

- config.rs: DeploymentConfig, CanonicalRoutingConfig, env-config
  functions, and process-id resolution
- solana.rs: Solana wallet verification, box burn, stable hash, and
  message signing utilities
- util.rs: Kernel type helpers, path utilities, event store error
  codes, and avatar generation

Also:
- Move core_burn_transaction test to solana.rs with the extracted code
- Move test-only config re-exports behind #[cfg(test)]
- Remove unused imports from main.rs (base64, ed25519, rand, CStr)
- Fix community_art_tests.rs base64 import
- Fix duplicate ai_publication tests after rebase
- Bump version to 0.0.330

main.rs: 72139 -> 72090 lines
