+++ v2/orchestrator-rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cosyworld-orchestrator"
-version = "0.0.329"
+version = "0.0.330"
 edition = "2021"
 publish = false
