+++ v2/orchestrator-rust/src/community_art_tests.rs
@@ -1,7 +1,7 @@
 use super::*;
 use crate::media_recipes::media_verdict::MEDIA_BRIEF_CONSTRAINT_LIMIT;
 use axum::{response::IntoResponse as _, routing::post, Router};
-use base64::Engine as _;
+use base64::{engine::general_purpose::STANDARD as BASE64_STANDARD, Engine as _};
 use image::{DynamicImage, ImageBuffer, ImageFormat, Rgba};
 use sha2::{Digest as _, Sha256};
 use std::{
